Consolidate GA helpers into selection utils
