+++ seq_mc_test.sh
@@ -109,13 +109,13 @@ EOS_PATH="$ITERATION_DIR/SIMULATIONS/EOS/eos_results.csv"
 DIFF_PATH="$ITERATION_DIR/SIMULATIONS/DIFF/diffusivities.csv"
 
 if [[ $ITER -eq 0 ]]; then
-    python /home/zl4808/PROJECTS/MODEL_COMPARISON/generate_labels.py \
+    python -m al_pipeline.labels.generate_labels \
         --eos_path "/scratch/gpfs/zl4808/PROJECTS/MODEL_COMPARISON/$MODEL/SIMULATIONS/EOS/eos_results.csv" \
         --diff_path "/scratch/gpfs/zl4808/PROJECTS/MODEL_COMPARISON/$MODEL/SIMULATIONS/DIFF/diffusivities.csv" \
         --output_path "$ITERATION_DIR/labels_gen$ITER.csv" \
         --iter $ITER
 else
-    python /home/zl4808/PROJECTS/MODEL_COMPARISON/generate_labels.py \
+    python -m al_pipeline.labels.generate_labels \
         --eos_path "$EOS_PATH" \
         --diff_path "$DIFF_PATH" \
         --output_path "$ITERATION_DIR/labels_gen$ITER.csv" \
@@ -149,7 +149,7 @@ python /home/zl4808/PROJECTS/MODEL_COMPARISON/train_gpr_multitask_mc.py \
 # -------------------------------
 echo "Generating parent sequences for iteration $ITER..."
 if [[ $ITER -eq 0 ]]; then
-    python /home/zl4808/PROJECTS/MODEL_COMPARISON/generate_parents.py \
+    python -m al_pipeline.selection.generate_parents \
         --features_path "$ITERATION_DIR/features_gen${ITER}_NORM_${EHVI}_${EXPLORE}_${TRANSFORM}_MC.csv" \
         --labels_path "$ITERATION_DIR/labels_gen${ITER}_NORM_${EHVI}_${EXPLORE}_${TRANSFORM}_MC.csv" \
         --output_path "$BASE_DIR" \
@@ -167,7 +167,7 @@ else
     cp "$SEQ_SIM_FILE" "$CURR_SEQ_FILE"
     SEQ_FILE="$ITERATION_DIR/seq_gen$ITER.txt"
     cat "$PREV_SEQ_FILE" "$CURR_SEQ_FILE" > "$SEQ_FILE"
-    python /home/zl4808/PROJECTS/MODEL_COMPARISON/generate_parents.py \
+    python -m al_pipeline.selection.generate_parents \
         --features_path "$ITERATION_DIR/features_gen${ITER}_NORM_${EHVI}_${EXPLORE}_${TRANSFORM}_MC.csv" \
         --labels_path "$ITERATION_DIR/labels_gen${ITER}_NORM_${EHVI}_${EXPLORE}_${TRANSFORM}_MC.csv" \
         --output_path "$BASE_DIR" \
@@ -188,7 +188,7 @@ cp "${BASE_DIR}/sequences_parent_TEMP_${EHVI}_${EXPLORE}_${TRANSFORM}_MC.txt" "$
 # Calculate Normalization Stats
 # -------------------------------
 echo "Calculating normalization statistics..."
-python /home/zl4808/PROJECTS/MODEL_COMPARISON/calculate_normalization_stats.py \
+python -m al_pipeline.features.calculate_normalization_stats \
     --features_file "$ITERATION_DIR/features_gen$ITER.csv" \
     --output_file "$ITERATION_DIR/normalization_stats.json"
